revert disksizer extraction
